RSA warning
/**
 *  RSA第三方警告 已修改
 - WARN  | [iOS] xcodebuild:  RSA.m:275:49: warning: values of type 'OSStatus' should not be used as format arguments; add an explicit cast to 'int' instead [-Wformat]
 - WARN  | [iOS] xcodebuild:  RSA.m:331:49: warning: values of type 'OSStatus' should not be used as format arguments; add an explicit cast to 'int' instead [-Wformat]
 */

+++ bench_ios/bench/thd/RSA/RSA.m
@@ -272,7 +272,7 @@ static NSData *base64_decode(NSString *str){
                                    );
         }
 		if (status != 0) {
-			NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+			NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
 			ret = nil;
 			break;
 		}else{
@@ -328,7 +328,7 @@ static NSData *base64_decode(NSString *str){
 							   &outlen
 							   );
 		if (status != 0) {
-			NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+			NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
 			ret = nil;
 			break;
 		}else{
